Move contains_any_tensor into yolort.utils
