+++ yolort/data/__init__.py
@@ -1,5 +1,1 @@
 # Copyright (c) 2021, yolort team. All rights reserved.
-
-from ._helper import contains_any_tensor
-
-__all__ = ["contains_any_tensor"]
