+++ yolort/utils/__init__.py
@@ -1,6 +1,8 @@
 # Copyright (c) 2020, yolort team. All rights reserved.
 
-from typing import Callable, Dict, Mapping, Sequence, Union
+from typing import Any, Callable, Dict, Mapping, Sequence, Type, Union
+
+from torch import Tensor
 
 try:
     from torch.hub import load_state_dict_from_url
@@ -10,17 +12,17 @@ except ImportError:
 from .dependency import check_version
 from .hooks import FeatureExtractor
 from .image_utils import cv2_imshow, get_image_from_url, read_image_to_tensor
-from .update_module_state import convert_yolov5_to_yolort, load_from_ultralytics
+from .update_module_state import convert_yolov5_to_yolort
 from .visualizer import Visualizer
 
 
 __all__ = [
     "check_version",
+    "contains_any_tensor",
     "cv2_imshow",
     "get_image_from_url",
     "get_callable_dict",
     "convert_yolov5_to_yolort",
-    "load_from_ultralytics",
     "load_state_dict_from_url",
     "read_image_to_tensor",
     "FeatureExtractor",
@@ -39,3 +41,16 @@ def get_callable_dict(fn: Union[Callable, Mapping, Sequence]) -> Union[Dict, Map
         return {get_callable_name(f): f for f in fn}
     elif callable(fn):
         return {get_callable_name(fn): fn}
+
+
+def contains_any_tensor(value: Any, dtype: Type = Tensor) -> bool:
+    """
+    Determine whether or not a list contains any Type
+    """
+    if isinstance(value, dtype):
+        return True
+    if isinstance(value, (list, tuple)):
+        return any(contains_any_tensor(v, dtype=dtype) for v in value)
+    elif isinstance(value, dict):
+        return any(contains_any_tensor(v, dtype=dtype) for v in value.values())
+    return False
